requirements/prod-requirements.txt
asgiref==3.7.2
Django==4.2.2django-cors-headers==3.14.0
django-countries==7.5.1
django-filter==23.1django-jazzmin==2.6.0
django-mysql==4.10.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
phonenumbers==8.13.8
protobuf==3.20.3
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4typing_extensions==4.6.3
tzdata==2023.3
drf-writable-nested==0.7.0
